refacto(web): #17 - change src folder structure
