+++ web/src/components/App.tsx
@@ -1,0 +1,14 @@
+import React from 'react';
+
+import BuilderPage from "../pages/BuilderPage";
+
+function App() {
+  return (
+    <>
+        <BuilderPage />
+    </>
+
+  );
+}
+
+export default App;
