+++ web/src/interfaces/ResourceNodeData.ts
@@ -1,0 +1,8 @@
+
+/**
+ * A Terraform Resource as React Flow node data
+ */
+export default interface ResourceNodeData {
+    type: string
+    label: string
+}
