+++ web/src/pages/BuilderPage/index.css
@@ -1,0 +1,70 @@
+.App {
+  text-align: center;
+}
+
+.App-logo {
+  height: 40vmin;
+  pointer-events: none;
+}
+
+@media (prefers-reduced-motion: no-preference) {
+  .App-logo {
+    animation: App-logo-spin infinite 20s linear;
+  }
+}
+
+.App-header {
+  background-color: #282c34;
+  min-height: 100vh;
+  display: flex;
+  flex-direction: column;
+  align-items: center;
+  justify-content: center;
+  font-size: calc(10px + 2vmin);
+  color: white;
+}
+
+.App-link {
+  color: #61dafb;
+}
+
+@keyframes App-logo-spin {
+  from {
+    transform: rotate(0deg);
+  }
+  to {
+    transform: rotate(360deg);
+  }
+}
+
+.wrapper {
+  display: grid;
+  grid-template-columns: repeat(6, 1fr);
+  grid-auto-rows: minmax(120px, auto);
+  height: 100%;
+  width: 100%;
+}
+.one {
+  grid-column: 1;
+  grid-row: 1/10;
+  /*background-color: red;*/
+  opacity: 50%;
+}
+.two {
+  grid-column: 2/6;
+  grid-row: 1 / 10;
+  background-color: blue;
+  opacity: 50%;
+}
+.three {
+  grid-column: 6;
+  grid-row: 1/10;
+  background-color: green;
+  opacity: 50%;
+}
+.four {
+  grid-column: 2/6;
+  grid-row: 9 / 10;
+  background-color: yellow;
+  opacity: 50%;
+}
